Final Major Code Upgrade
Changes:

  [ ADDED ]

  * Added program and developer documentation
  * Added a debugger to debug the program
  * Added Debug Builds and Release builds scripts
  * Added ASCII ART format for program info.
  * Added tree view in file manager
  * Added copy/move function in file manager.

  [ IMPROVEMENTS ]

  * Improved UpdateCredentials program
  * Improved program scripts
  * Improved the ErrorHandler program
  * Improved policy enforcement efficiency.
  * Improved the Time of Day detection logic.
  * Improved setup code significantly
  * Improved the help manuals
  * Improved LoginAPI, AddUser and UpdateCredentials.
  * Improved memory usage by removing unwanted classes.
    in MainMenu program.

  [ REMOVED ]

  * Removed Messenger program.
  * Remove ChangePassword program.
  * Removed Add user command from guest user interface.
  * Fixed bug where Reset didn't work until the program
    was restarted.

  [ FIXED ]

  * Fixed documentation format to latest specification.
  * Fixed memory leak issues.
  * Fix bug where user files are re-created if deleted.

  [ CHANGED ]

  * Changed database table name from FSAD to FCAD
    (Fractal Credential Account Database)

+++ Source/API/Tools/FileManager/FileManager.java
@@ -1,3 +1,19 @@
+/*
+ *****************************************************
+ *                                                   *
+ * ! IMPORTANT ! DO NOT DELETE COMMENT ! IMPORTANT ! *
+ *                                                   *
+ *****************************************************
+ *                                                   *
+ *            THIS CODE IS RELEASE READY.            *
+ *                                                   *
+ *      THIS CODE HAS BEEN TESTED, REVIEWED AND      *
+ *      REVISED. THIS CODE HAS NO KNOWN ISSUES,      *
+ *      HENCE IT IS CONSIDERED AS RELEASE READY      *
+ *                                                   *
+ *****************************************************
+ */
+ 
 package API.Tools.FileManager;
 
 import java.io.*;
@@ -5,6 +21,21 @@ import java.io.*;
 import API.*;
 import API.Anima.*;
 
+
+/**
+* A class to implement the FileManager functionality
+* <BR>
+* <pre>
+* ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+* |            TECHNICAL DETAILS            |
+* ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+* | Class ID    :  BA3-Mosaic-FileMgr-FMGR  |
+* | Class Name  :  FileManager              |
+* | Since       :  0.0.1, 17-August-2020    |
+* | Updated on  :  0.2.1, 04-October-2020   |
+* ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+* </pre>
+*/
 public final class FileManager
 {
 	private boolean SB=false;
@@ -23,7 +54,13 @@ public final class FileManager
 	API.Tools.Download dl=new API.Tools.Download();
 	API.SHA256 sha=new API.SHA256();
 	
-	//take secureboot, username and admin values
+	/**
+	* Constructor to initialize the SecureBoot variable, the current user's username and name.
+	*
+	* @param SecureBoot : Receive the secure boot status from the previous program.
+	* @param Username   : Receive the username of the account from the previous program.
+	* @param name       : Receive the name of the user to make the user path and username readable.
+	*/
 	public FileManager(boolean SecureBoot, String Username, String name)
 	{
 		if(SecureBoot==false)
@@ -33,6 +70,12 @@ public final class FileManager
 		Name=name;
 	}
 	
+	
+	/**
+	* Method to access the encapsulated FileManager publicly.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/
 	public void FileManagerScript()throws Exception
 	{
 		API.policyEnforce pe=new API.policyEnforce("filemanager");
@@ -66,6 +109,11 @@ public final class FileManager
 	* decrypt file
 	*/
 	
+	/**
+	* Method which actually implements the file manager functionality.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution.
+	*/	
 	private void FileManager()throws Exception
 	{
 		System.gc();
@@ -92,6 +140,10 @@ public final class FileManager
 							listFiles();
 							break;
 							
+				case "tree":
+							TreeView();
+							break;				
+							
 				case "change dir":
 							changeDir();
 							break;
@@ -103,10 +155,6 @@ public final class FileManager
 				case "delete":
 							del();
 							break;
-							
-				case "rename":
-							rename();
-							break;
 				
 				case "encrypt": 
 							enc.encr(User, CurDir);
@@ -117,15 +165,27 @@ public final class FileManager
 							break;
 				
 				case "read file":
-							rF.userReadFile(User, CurDir);
+							rF.userReadFile(CurDir);
 							break;
 				
 				case "edit file": 
-							te.editScript(User, CurDir);
+							te.editScript(CurDir);
 							break;
 							
 				case "download":
 							dl.downloadFile(User, CurDir);
+							break;
+							
+				case "copy":
+							copy_move_frontend(false);
+							break;
+				
+				case "move":
+							copy_move_frontend(true);
+							break;
+							
+				case "rename":
+							rename();
 							break;
 				
 				case "": 
@@ -141,6 +201,11 @@ public final class FileManager
 	
 	// --- make a common interface to check if the directory exists or not. --- //
 	
+	/**
+	* Method which confirms the identity of the user currently logged in.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution.
+	*/
 	private boolean confirmIdentity()throws Exception
 	{
 		String Password, Key;
@@ -149,10 +214,16 @@ public final class FileManager
 		System.out.println("Username: "+Name);
 		Password=sha.encodedString(String.valueOf(console.readPassword("Password: ")));
 		Key=sha.encodedString(String.valueOf(console.readPassword("Security Key: ")));
-		API.Anima.LoginAPI login=new API.Anima.LoginAPI(SB, User, Password, Key);
+		API.Anima.LoginAPI login=new API.Anima.LoginAPI(User, Password, Key);
 		return login.status();
 	}
 	
+	
+	/**
+	* Method to implement the functionality of changing the current directory.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/	
 	private void changeDir()throws Exception
 	{
 		String tPath=console.readLine("Enter the directory name: ");
@@ -170,21 +241,70 @@ public final class FileManager
 		return;
 	}
 	
-	private void prevDir()
+	
+	/**
+	* A sub-method within changeDir() implemented to change to the previous directory.
+	*
+	* NOTE: Program will not allow the directory to access levels above the user home path.
+	*/
+	private void prevDir()throws Exception
 	{
 		CurDir=CurDir.substring(0, CurDir.length()-1);
 		CurDir=CurDir.replace(CurDir.substring(CurDir.lastIndexOf('/'), CurDir.length()), "/");
 		if(CurDir.equals("./Users/"))
 		{
-			System.out.println("[ WARNING ] : File Management cannot be out of user's home path. Reverting to default user's home path.");
+			System.out.println("[ WARNING ] : File Management cannot be above user's home path. Reverting to user's home path.");
 			CurDir="./Users/"+User+"/";
 		}
 		return;
 	}
 	
+	/**
+	* Method to print the tree view of the given path.
+	*
+	* @throws Exception : Throws any exception caught during runtime/execution.
+	*/
+	private void TreeHelper(int indent, File file)
+	{
+		System.out.print("|");
+		for (int i = 0; i < indent; ++i)
+		  System.out.print('-');
+		System.out.println(file.getName().replace(User,Name + " [ HOME PATH ROOT ]"));
+		if (file.isDirectory()) {
+		  File[] files = file.listFiles();
+		  for (int i = 0; i < files.length; ++i)
+			TreeHelper(indent + 2, files[i]);
+		}
+	}
+	
+	/**
+	* Method to view the files in a tree hierarchical view.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution.
+	*/
+	private void TreeView()throws Exception
+	{
+		try
+		{
+			File tree=new File(CurDir);
+			System.out.println("\n--- [ TREE VIEW ] ---\n");
+			TreeHelper(0, tree);
+			System.out.println();
+			return;
+		}
+		catch(Exception E)
+		{
+			E.printStackTrace();
+		}
+	}
+
+	/**
+	* Method to implement functionality to list the files in a given directory.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/
 	private void listFiles()throws Exception
 	{
-		
 		//String format = "%1$-60s|%2$-50s|%3$-20s\n";
 		String format = "%1$-50s|%2$-20s\n";
 		if(checkFile(CurDir)==true)
@@ -204,6 +324,11 @@ public final class FileManager
 		return;
 	}
 	
+	/**
+	* A common method used to check if a file exists in a given directory or not.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/
 	private boolean checkFile(String fName)throws Exception
 	{
 		f=new File(fName);
@@ -212,7 +337,12 @@ public final class FileManager
 		return true;
 	}
 	
-	private void makeDir()
+	/**
+	* Method to implement the functionality to create a new directory
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/
+	private void makeDir()throws Exception
 	{
 		try
 		{
@@ -234,7 +364,12 @@ public final class FileManager
 		}
 	}
 	
-	private void del()
+	/**
+	* Method to implement the functionality to delete a given file or directory from a specified path.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/
+	private void del()throws Exception
 	{
 		try
 		{
@@ -243,7 +378,14 @@ public final class FileManager
 			if(checkFile(delFile)==true)
 			{
 				f=new File(delFile);
-				f.delete();
+				if(f.isDirectory()==true)
+				{
+					delHelper(f);
+				}
+				else
+				{
+					f.delete();
+				}
 			}
 			else
 				System.out.println("[ ERROR ] : The specified file/directory does not exist.");
@@ -256,6 +398,26 @@ public final class FileManager
 		}
 	}
 	
+	/**
+	* A sub-method which helps in deleting directories within the directory to be deleted.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/
+	private void delHelper(File delfile)throws Exception
+	{
+		if (delfile.listFiles() != null) 
+		{
+			for (File fock : delfile.listFiles()) 
+				delHelper(fock);
+		}
+		delfile.delete();
+	}
+	
+	/**
+	* A method to implement the functionality to rename a given file or folder.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/
 	private void rename()
 	{
 		try
@@ -275,6 +437,72 @@ public final class FileManager
 		catch (Exception E)
 		{
 			//troubleshooting tips here
+			E.printStackTrace();
+		}
+	}
+	
+	/**
+	* Method to copy or move the files in file manager.
+	*
+	* @throws Exception  : Throws any exception caught during runtime/execution
+	*/
+	private void copy_move_frontend(boolean move)throws Exception
+	{
+		try
+		{
+			String source=console.readLine("Enter the source file/directory name: ");
+			String destination=console.readLine("Enter the destination file/directory name: ");
+			File s=new File(CurDir+source);
+			File d=new File(CurDir+destination);
+			copy_move_helper(s, d);
+			if(move==true)
+			{
+				delHelper(s);
+			}
+			return;
+		}
+		catch(Exception E)
+		{
+			E.printStackTrace();
+		}
+	}
+	
+	/**
+	* Method which will help in copying or moving the files.
+	*
+	* @param src : pass the source file to the implementation .
+	* @param dest : pass the destination file to the implementation.
+	* @throws Exception : Throws any exception caught during runtime/execution
+	*/
+	private void copy_move_helper( File src, File dest ) throws Exception 
+	{
+		try
+		{
+			if( src.isDirectory() )
+			{
+				dest.mkdirs();
+				for( File sourceChild : src.listFiles() ) 
+				{
+					File destChild = new File( dest, sourceChild.getName() );
+					copy_move_helper( sourceChild, destChild );
+				}
+			} 
+			else
+			{
+				InputStream in = new FileInputStream( src );
+				OutputStream out = new FileOutputStream( dest );
+				byte[] buf = new byte[1024];
+				int len;
+				while ((len = in.read(buf)) > 0) 
+				{
+					out.write(buf, 0, len);
+				}
+				in.close();
+				out.close();
+			}
+		}
+		catch(Exception E)
+		{
 			E.printStackTrace();
 		}
 	}
